// src/interfaces/Filmografia.ts
export interface Filmografia {
  uid: string
  slug: string
  year: number
  pais: string
  titulo: string
  title: string
  productora: string
  filmada_en: string
  direccion: string
  guion: string
  fotografia: string
  musica: string
  duracion: string
  protagonistas: string
  sipnopsis: string
  synopsis: string
  cartel: string
  cartel_url: string
}